demo_vulnerable_app/requirements.txt
flask==2.3.0
requests==2.31.0
pyyaml==6.0
